[nrf fromlist] Bluetooth: Controller: Move BT_HCI_VS_FATAL_ERROR impl to hci_vs_err.c
Move the CONFIG_BT_HCI_VS_FATAL_ERROR implementation out of the
monolithic hci.c and into a new dedicated file
subsys/bluetooth/controller/hci/hci_vs_err.c, with public declarations
in subsys/bluetooth/controller/include/hci_vs_err.h.

This makes hci_vs_err_assert() and hci_vs_err_stack_frame() available
for out-of-tree controller implementations such as SoftDevice
Controller in nRF Connect SDK, without requiring inclusion of the full
Zephyr HCI controller implementation.

Upstream PR #: 114829
